requirements.txt
pyrogram
html_telegraph_poster
TgCrypto
psutil
pymongo[srv]
aiohttp
python-dotenv
hachoir
anitopy
pyromod
asyncio
